add the cloud cover data to the system

+++ app/src/main/java/io/github/vinge1718/weather/models/Clouds.java
@@ -8,7 +8,7 @@ public class Clouds {
 
     @SerializedName("all")
     @Expose
-    private Integer all;
+    private Long all;
 
     /**
      * No args constructor for use in serialization
@@ -21,16 +21,16 @@ public class Clouds {
      * 
      * @param all
      */
-    public Clouds(Integer all) {
+    public Clouds(Long all) {
         super();
         this.all = all;
     }
 
-    public Integer getAll() {
+    public Long getAll() {
         return all;
     }
 
-    public void setAll(Integer all) {
+    public void setAll(Long all) {
         this.all = all;
     }
